app/requirements.txt
fastapi
uvicorn
mediapipe
opencv-python
deepface
insightface
numpy
pillow
python-multipart
pydantic
tf-keras-vis
tf-keras
onnxruntime
tensorflow
